finalized account creation, management, and deletion

+++ app/controllers/users_controller.rb
@@ -1,7 +1,7 @@
 class UsersController < ApplicationController
 
-	skip_before_action :authorize, only: [:new, :create, :confirm]
-	skip_before_action :current_user, only: [:new, :create, :confirm]
+	skip_before_action :authorize, only: [:new, :create, :confirm, :deny]
+	skip_before_action :current_user, only: [:new, :create, :confirm, :deny]
 	before_action :confirm_user, only: [:edit, :show, :update, :destroy]
 
 	def new
@@ -9,12 +9,16 @@ class UsersController < ApplicationController
 	end
 
 	def create
-		existing = User.find_by_email(params[:user][:email])
+		existing_email = User.find_by_email(params[:user][:email])
+		existing_uname = User.find_by_username(params[:user][:username])
 
-		if existing != nil
+		if existing_email != nil
 			flash[:danger] = 'Sorry, that email is already taken.'
 			redirect_to new_user_path
-		elsif existing == nil
+		elsif existing_uname != nil
+			flash[:danger] = 'Sorry, that username is already taken.'
+			redirect_to new_user_path
+		elsif existing_email == nil && existing_uname == nil
 			@user = User.create(user_params)
 
 			if @user.save
@@ -37,10 +41,26 @@ class UsersController < ApplicationController
 			@user.account_create_confirmed_at = Time.zone.now
 
 			if @user.save
+				cookies[:auth_token] = @user.auth_token
 				params[:account_create_token] = @user.account_create_token
 			else
 				flash[:danger] = 'We\'re sorry, there was an error creating your account.'
 			end
+		end
+	end
+
+	def deny
+		@user = User.find_by_account_create_token!(params[:id])
+
+		if @user.account_create_confirmed_at == nil
+			if @user.destroy
+				flash[:info] = 'The account associated with this email has been deleted. We apologize for the error.'
+				redirect_to root_path
+			else
+				flash[:danger] = 'There was an error deleting the account associated with your email. Please contact support.'
+			end
+		else
+			flash.now[:info] = 'The account associated with this email has already been confirmed. It can only be deleted by the authorized user.'
 		end
 	end
 
@@ -72,13 +92,24 @@ class UsersController < ApplicationController
 	end
 
 	def destroy
-		if User.destroy_all(id: params[:id])
-			flash[:success] = 'You have successfully deleted your account.'
-			session[:id] = nil
-			redirect_to root_path
-		else
-			flash[:danger] = 'Your account was not deleted.'
-			render :edit
+		@user = User.find(params[:id])
+
+		if delete_user_params[:current_password].length == 0
+			flash.now[:danger] = 'Please enter your password to delete your account.'
+			render :show
+		elsif @user.authenticate(delete_user_params[:current_password]) != @user
+			flash.now[:danger] = 'The password you entered does not match the current account password.'
+			render :show
+		elsif @user.authenticate(delete_user_params[:current_password]) == @user
+			if User.destroy_all(id: params[:id])
+				flash[:success] = 'You have successfully deleted your account.'
+				session[:id] = nil
+				cookies[:auth_token] = nil
+				redirect_to root_path
+			else
+				flash[:danger] = 'Your account was not deleted.'
+				render :edit
+			end
 		end
 	end
 
@@ -92,4 +123,7 @@ private
 		params.require(:user).permit(:first_name, :last_name, :email, :username, :current_password, :password, :password_confirmation)
 	end
 
+	def delete_user_params
+		params.require(:user).permit(:current_password)
+	end
 end
